LayoutTests/accessibility/mac/select-text/select-text-7-expected.txt

The slow brown fox jumps over the lazy dog.

TEXT2: The quick brown fox jumps over the lazy dog. The quick brown fox jumps over the lazy dog.

This tests the ability to select and replace text with respect to selection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 'slow'
PASS selection is 'slow'
PASS successfullyParsed is true

TEST COMPLETE
